requirements.txt
numpy>=1.24
pandas>=2.2.2
scipy>=1.11
statsmodels>=0.14.0
scikit-learn>=1.3
matplotlib>=3.8
pyyaml>=6.0
numba>=0.59
